requirements.txt
Click
psycopg2-binary
tabulate